Fixed #56: add --dry-run command line parameter.
This commit adds the --dry-run parameter to validate the input data and print
an output summary even in the presence of errors. To this end, the --output
parameter is also introduced to give an option to display the summary as JSON
or in a more user-friendly text format. The --debug parameter will repeat the
full list of errors when there are any at the end of the run. Errors are now
printed to stderr and may be surpressed using shell redirection.