Add --dry-run command line parameter.
This commit fixes issue #56 by adding the --dry-run parameter to validate the
input data and print an output summary even in the presence of errors. To this
end, the --output parameter is also introduced to give an option to display the
summary as JSON or in a more user-friendly text format. Data should not be
loaded if there are errors or if this command is invoked with --dry-run.